requirements.txt
az==0.1.0.dev1
azure-common==1.1.25
azure-core==1.4.0
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.6
azure-nspkg==3.0.2
azure-storage-blob==12.3.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.9.2
Flask==1.1.1
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
httplib2==0.18.1
idna==2.9
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
msrest==0.6.13
oauthlib==3.1.0
pycparser==2.20
PyJWT==1.7.1
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
SQLAlchemy==1.3.15
twilio==6.41.0
urllib3==1.25.9
Werkzeug==0.16.1
